refactor(tests): share research runner and centralize features
