+++ tests/server_contracts/internal/research/stream_write_steps_test.go
@@ -1,4 +1,4 @@
-package topicresearch_test
+package research_test
 
 import (
 	"context"
