+++ tests/server_contracts/internal/runner/compose.go
@@ -1,4 +1,4 @@
-package main
+package runner
 
 import (
 	"context"
